+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # NeON
 
+## 2025-04-30: ???
+
+- Added `write()` builtin to write text into a file
+
 ## 2025-04-14: 1.13.1
 
 - Fixed code warnings